requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Flask==2.0.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
outcome==1.1.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
selenium==4.1.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
twilio==7.5.1
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.3
wsproto==1.0.0
zipp==3.7.0
